refactor: Modify Sudoku solver to count solutions
The Sudoku solver in the `sudoku.py` file has been modified to include a new function `count_solutions()`. This function uses a modified backtracking algorithm to count the number of solutions for a given Sudoku board. It recursively tries different numbers in empty spots and backtracks if a solution is found. The function also includes a base case to stop counting if more than one solution is found.

Additionally, a new function `has_unique_solution()` has been added to verify if a Sudoku board has exactly one solution. This function makes a copy of the board and uses the `count_solutions()` function to determine the number of solutions. If exactly one solution is found, the board is considered to have a unique solution.

These changes improve the functionality of the Sudoku solver and provide a way to check if a board has a unique solution.